Move parseAttributes() to lib/attributes
